refactor: move page implementation to scenes
